+++ proptest/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Bug Fixes
 
 - Fixed a panic when sampling from a single-point inclusive float range like `0.0..=0.0`. ([\#479](https://github.com/proptest-rs/proptest/issues/479))
+- [Soundness] Fixes data race on static mut between assigning thread and accessing thread ([\#648](https://github.com/proptest-rs/proptest/issues/648))
 
 ### Other Notes